refactor: ResourceMapper과 ServiceMapper 클래스 분리
