+++ src/main/java/webserver/controller/RequestController.java
@@ -1,7 +1,7 @@
 package webserver.controller;
 
 import webserver.http.request.HttpRequest;
-import webserver.http.HttpResponse;
+import webserver.http.response.HttpResponse;
 
 public abstract class RequestController {
     private static final String INDEX_HTML_PATH = "/index.html";
