+++ src/main/java/webserver/controller/resource/CssController.java
@@ -1,7 +1,7 @@
 package webserver.controller.resource;
 
 import webserver.http.request.HttpRequest;
-import webserver.http.HttpResponse;
+import webserver.http.response.HttpResponse;
 import webserver.controller.RequestController;
 
 public class CssController extends RequestController {
